refactor vector search helpers
