+++ tests/helpers/vectorSearchPage.test.js
@@ -5,46 +5,6 @@ vi.doMock("../../src/helpers/domReady.js", () => ({
   onDomReady: vi.fn()
 }));
 
-/**
- * Unit tests for selectMatches helper in vectorSearchPage.
- */
-
-describe("selectMatches", () => {
-  it("returns only the top match when drop off exceeds threshold", async () => {
-    const { selectMatches } = await import("../../src/helpers/vectorSearchPage.js");
-    const strong = [
-      { id: "1", score: 0.95 },
-      { id: "2", score: 0.4 },
-      { id: "3", score: 0.39 }
-    ];
-    const result = selectMatches(strong, []);
-    expect(result).toHaveLength(1);
-    expect(result[0].id).toBe("1");
-  });
-
-  it("returns all strong matches when drop off is small", async () => {
-    const { selectMatches } = await import("../../src/helpers/vectorSearchPage.js");
-    const strong = [
-      { id: "1", score: 0.8 },
-      { id: "2", score: 0.7 }
-    ];
-    const result = selectMatches(strong, []);
-    expect(result).toEqual(strong);
-  });
-
-  it("uses weak matches when no strong matches exist", async () => {
-    const { selectMatches } = await import("../../src/helpers/vectorSearchPage.js");
-    const weak = [
-      { id: "a", score: 0.5 },
-      { id: "b", score: 0.4 },
-      { id: "c", score: 0.3 },
-      { id: "d", score: 0.2 }
-    ];
-    const result = selectMatches([], weak);
-    expect(result).toEqual(weak.slice(0, 3));
-  });
-});
-
 describe("vector search page integration", () => {
   it("passes selected tag to findMatches", async () => {
     const findMatches = vi.fn().mockResolvedValue([]);
@@ -58,9 +18,8 @@ describe("vector search page integration", () => {
       }
     }));
 
-    const { handleSearch, init, __setExtractor } = await import(
-      "../../src/helpers/vectorSearchPage.js"
-    );
+    const { handleSearch, init } = await import("../../src/helpers/vectorSearchPage.js");
+    const { __setExtractor } = await import("../../src/helpers/api/vectorSearchPage.js");
 
     __setExtractor(async () => ({ data: [0, 0, 0] }));
 
@@ -145,9 +104,8 @@ describe("search result message styling", () => {
       DATA_DIR: "./"
     }));
 
-    const { handleSearch, init, __setExtractor } = await import(
-      "../../src/helpers/vectorSearchPage.js"
-    );
+    const { handleSearch, init } = await import("../../src/helpers/vectorSearchPage.js");
+    const { __setExtractor } = await import("../../src/helpers/api/vectorSearchPage.js");
 
     __setExtractor(async () => ({ data: [0, 0, 0] }));
 
@@ -196,9 +154,8 @@ describe("search result message styling", () => {
       DATA_DIR: "./"
     }));
 
-    const { handleSearch, init, __setExtractor } = await import(
-      "../../src/helpers/vectorSearchPage.js"
-    );
+    const { handleSearch, init } = await import("../../src/helpers/vectorSearchPage.js");
+    const { __setExtractor } = await import("../../src/helpers/api/vectorSearchPage.js");
 
     __setExtractor(async () => ({ data: [0, 0, 0] }));
 
@@ -247,9 +204,8 @@ describe("search result message styling", () => {
       DATA_DIR: "./"
     }));
 
-    const { handleSearch, init, __setExtractor } = await import(
-      "../../src/helpers/vectorSearchPage.js"
-    );
+    const { handleSearch, init } = await import("../../src/helpers/vectorSearchPage.js");
+    const { __setExtractor } = await import("../../src/helpers/api/vectorSearchPage.js");
 
     __setExtractor(async () => ({ data: [0, 0, 0] }));
 
@@ -313,9 +269,8 @@ describe("snippet highlighting", () => {
       DATA_DIR: "./"
     }));
 
-    const { handleSearch, init, __setExtractor } = await import(
-      "../../src/helpers/vectorSearchPage.js"
-    );
+    const { handleSearch, init } = await import("../../src/helpers/vectorSearchPage.js");
+    const { __setExtractor } = await import("../../src/helpers/api/vectorSearchPage.js");
 
     __setExtractor(async () => ({ data: [0, 0, 0] }));
 
@@ -388,9 +343,8 @@ describe("synonym expansion", () => {
       DATA_DIR: "./"
     }));
 
-    const { handleSearch, init, __setExtractor } = await import(
-      "../../src/helpers/vectorSearchPage.js"
-    );
+    const { handleSearch, init } = await import("../../src/helpers/vectorSearchPage.js");
+    const { __setExtractor } = await import("../../src/helpers/api/vectorSearchPage.js");
 
     let captured = "";
     __setExtractor(async (text) => {
